requirement.txt
bootstrap4==0.1.0
Django==4.2.13
django-allauth==0.63.3
django-bootstrap4==24.3
django-chunked-upload==2.0.0
django-widget-tweaks==1.5.0
psycopg2-binary==2.9.9
